Rename `no-fn-reference-in-iterator` to `no-array-callback-reference`
