+++ index.js
@@ -31,9 +31,9 @@ module.exports = {
 				'unicorn/import-style': 'error',
 				'unicorn/new-for-builtins': 'error',
 				'unicorn/no-abusive-eslint-disable': 'error',
+				'unicorn/no-array-callback-reference': 'error',
 				'unicorn/no-array-instanceof': 'error',
 				'unicorn/no-console-spaces': 'error',
-				'unicorn/no-fn-reference-in-iterator': 'error',
 				'unicorn/no-for-loop': 'error',
 				'unicorn/no-hex-escape': 'error',
 				'unicorn/no-keyword-prefix': 'off',
